+++ packages/editor/web/src/RichContentEditor/utils/pasting/rgbToHex.js
